ml-pipeline-requirements.txt
pandas
sqlalchemy
psycopg[binary]
scikit-learn
numpy
python-dotenv
